+++ tracecat/expressions/functions.py
@@ -12,7 +12,7 @@ import orjson
 from jsonpath_ng.exceptions import JsonPathParserError
 
 from tracecat.expressions.shared import ExprContext
-from tracecat.expressions.validators import is_iterable
+from tracecat.expressions.validation import is_iterable
 from tracecat.logging import logger
 from tracecat.types.exceptions import TracecatExpressionError
